backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0  # rest framework
markdown==3.8  # Markdown support for the browsable API
django-filter==25.1  # Filtering support
django-cors-headers==4.7.0  # CORS support
psycopg2-binary==2.9.10  # PostgreSQL adapter (for production)
openai==1.84.0  # OpenAI API client
httpx==0.28.1  # async HTTP client for external APIs
structlog==25.4.0  # structured logging
